+++ HISTORY.rst
@@ -19,6 +19,7 @@ History
 * Fix map of array of primitive handling
 * Escape chars are no longer lost in embedded AVRO schema
 * Mandatory record fields are now checked for a value
+* Embedded schema now matches original one
 
 0.3.3 (2022-01-26)
 ^^^^^^^^^^^^^^^^^^